\name{twInverseTutorial-package}
\alias{twInverseTutorial-package}
\title{Hands-on-session on inverse modelling}
\description{Hands on session for course on Maximum Likelihood and Bayesian inverse parameter estimation}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}

\keyword{ package }


